Fixed validation errors
- A modelform's clean method sets a flag to validate for uniqueness,
which means submitting a share form with a correct sharecode would
actually fail validation as "a model already exists with that
sharecode". Since this form is just for lookup and not for saving new
models from it, we can override the clean() method to diable the unique
validation flag.
- The jquery .text() method escapes it's contents, so combining it with
the htmlEscape() function was double escaping user inputs